prevent NPE when BuildData is missing from the build
It's possible that getAction() will return a null pointer, if there are
no such Actions of the requested class. This typically wouldn't occur as
we normally expect the git-plugin to store this data. However, it may
not in some cases such as if the plugin crashed.

Additionally, future plugin versions may stop storing BuildData in favor
of another format (as the BuildData can become bloated with a growing
hash of branch/build pairs in the pathological case).

Make the FavoritingScmListener safe against such null pointers.

Add a new test which deletes the BuildData as soon as the job finishes,
to verify that a null build data doesn't cause a crash, and doesn't
result in a favorite. It's a bit hacky, since we're time dependent, so
we need to significantly decrease the times passed to Thread.sleep().

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>