Fix test to avoid possible infinite loop
When updating the test code for createAndRunPipeline, functionality was
added to allow attempting to remove the Git BuildData action as soon as
possible. This test code was flawed because in the case where the
BuildData was never added, it would loop forever. Instead, merge the
loop together into the same loop as the run.getResult(). This way, we
will keep attempting to remove the BuildData action until the job
finishes. This works because the Git BuildData action should be added
during the build, prior to it finishing.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>